Sneaker Store Database
I was tasked to imagine I have a store and create a database for some of the items I would be selling in that store. I created the table and inserted important columns such as sizes, price, color, and an ID for each item. I then chose to display the database in order by price and